(FIND-1528) uri-encode the string in a sane way; replacing code which has multiple bugs which cancel out by coincidence. Removes an unnecessary reduce and 3 pointless replace+encodeURIComponent calls.
Firstly, we want to encode the whole string, not just those three characters; otherwise '_' would give problems with QMS as in FIND-1524.
Secondly, the _.reduce takes the memo as the first argument, which is not the token. So the characters in the array don't even matter.

In practice, this previously worked because there was an odd number of elements in the array, which effectively works out to be equivalent to just `encodeURIComponent`on its own. However, if you add another character e.g. ' ' such that there's an even number of elements in the array, then no encoding is done.